refactor: share logger lifecycle without changing wandb semantics
